+++ src/services/review.service.js
@@ -1,6 +1,7 @@
-import { getTransactionsBySeller, getTransactionsByBuyer } from '@/services/transaction.service';
+import { getTransactionsBySeller, getTransactionsByBuyer, updateTransaction } from '@/services/transaction.service';
 import { getDisplayName } from '@/services/user.service';
-import { isEmptyObject } from '@/services/utils.service';
+import { getCurrentTimestamp, isEmptyObject } from '@/services/utils.service';
+import { Review } from '@/models/review.class';
 
 const getReviews = async (userId, role) => {
   const key = role === 'Buyer' ? 'buyerReview' : 'sellerReview';
@@ -21,14 +22,26 @@ const getReviews = async (userId, role) => {
 const getAggregatedRating = (reviews) => {
   const length = reviews.length;
   if (length) {
-    const total = reviews.map((ele) => { return ele.rating ? ele.rating : 0; }).reduce((a, b) => { return a + b, 0; });
+    const total = reviews.map((ele) => { return ele.rating ? ele.rating : 0; }).reduce((a, b) => { return a + b; });
     return (total / length);
   }
   return 0;
+};
+
+const updateBuyerReview = async (transactionId, rating, description) => {
+  const reviewPayload = new Review(rating, description, getCurrentTimestamp());
+  return await updateTransaction(transactionId, { buyerReview: { ...reviewPayload } });
+};
+
+const updateSellerReview = async (transactionId, rating, description) => {
+  const reviewPayload = new Review(rating, description, getCurrentTimestamp());
+  return await updateTransaction(transactionId, { sellerReview: { ...reviewPayload } });
 };
 
 
 export {
   getReviews,
   getAggregatedRating,
+  updateBuyerReview,
+  updateSellerReview,
 };
